refactor(cli): extract workflow report handler
